# internal/handlers/openapi.yaml
openapi: 3.1.0
info:
  title: whatsmeow-api
  version: 0.1.0
  description: |
    REST API on top of go.mau.fi/whatsmeow. Multi-instance WhatsApp
    number management with a web manager UI and a per-instance API
    for sending text + interactive buttons.

    Two auth surfaces:
    - **Manager** (env-stored password → session cookie) — `/admin/*` and `/admin/api/*`
    - **Per-instance** (`Authorization: Bearer sk_live_...`) — `/api/v1/*`
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /admin/logout:
    post:
      tags: [manager]
      summary: Invalidate the current session
      security: [{ sessionCookie: [] }]
      responses: { "200": { description: Logged out } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: "whatsmeow-api-key (sk_live_...)"
...